pyproject.toml
[project]
name = "network-security-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dagshub>=0.6.3",
    "dill>=0.4.0",
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "mlflow<3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pymongo[srv]>=4.14.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "streamlit>=1.49.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "dagshub>=0.6.3",
    "mlflow<3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pymongo[srv]>=4.14.0",
    "python-dotenv>=1.1.1",
]
